Add edit functionality to BookMarkButton
In mounted method it's call fetch function that try to retrieve data from api to understand if the current book is already marked as read or in reading by the current user. If yes, it loads the info and put them on screen and change the save click callback from add to edit, if not, the callback it will be add function and no info will be touched.